requirements.txt

# Core
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# PDF Processing
PyMuPDF==1.23.8
pypdf==3.17.4

# LLM & LangChain
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# API
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23

# UI
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4

# Metrics & Validation
scikit-learn==1.3.2
rouge-score==0.1.2
nltk==3.8.1

# Utils
python-dateutil==2.8.2
tqdm==4.66.1